requirements.txt
Django==4.0.3
pillow
slugify
python-slugify
django-jazzmin
django-ckeditor
django-import-export
